fixes loading secret
